+++ experiments/archive_model_builders.py
@@ -1,3 +1,4 @@
+# flake8: noqa
 import os
 from typing import Any, Dict
