+++ experiments/benchmark.py
@@ -1,12 +1,6 @@
-# flake8: noqa
-import logging
-import os
-import random
-
-import numpy as np
 import pandas as pd
-import torch
-from optuna_params import FIXED_PARAMS
+from model_evaluation import evaluate_model
+from param_space import FIXED_PARAMS
 
 from darts import TimeSeries
 from darts.datasets import (
@@ -21,91 +15,19 @@ from darts.models import (
     ARIMA,
     FFT,
     CatBoostModel,
-    DLinearModel,
-    LightGBMModel,
     LinearRegressionModel,
     NaiveSeasonal,
     NBEATSModel,
     NHiTSModel,
-    NLinearModel,
     Prophet,
     TCNModel,
-    XGBModel,
 )
-from darts.models.forecasting.forecasting_model import LocalForecastingModel
-
-logging.getLogger("pytorch_lightning").setLevel(logging.WARNING)
 
 
-def randommness(seed=0):
-    random.seed(seed)
-    torch.manual_seed(seed)
-    np.random.seed(seed)
-    torch.use_deterministic_algorithms(True)
-
-
-def convert_to_ts(ds):
+def convert_to_ts(ds: TimeSeries):
     return TimeSeries.from_times_and_values(
         pd.to_datetime(ds.time_index), ds.all_values(), columns=ds.components
     )
-
-
-def evaluate_model(
-    model_class,
-    dataset,
-    model_params=dict(),
-    metric=mae,
-    split=0.7,
-    past_cov=None,
-    future_cov=None,
-    **kwargs
-):
-
-    model_uses_cov = model_params.get(
-        "lags_future_covariates", False
-    ) or model_params.get("lags_past_covariates", False)
-    if model_uses_cov and not (past_cov or future_cov):
-        raise ValueError("model uses covariates, but none were provided")
-    if past_cov and len(dataset) != len(past_cov):
-        raise ValueError("past_cov and dataset must have the same length")
-    if future_cov and len(dataset) != len(future_cov):
-        raise ValueError("future_cov and dataset must have the same length")
-
-    model = model_class(**model_params)
-    num_test_points = 20
-    stride = int((1 - split) * len(dataset) / num_test_points)
-    stride = max(stride, 1)
-    retrain = True
-    if not isinstance(model, LocalForecastingModel):
-        retrain = False
-
-        train = dataset.split_after(split)[0]
-        train_past_cov = past_cov.split_after(split)[0] if past_cov else None
-        train_future_cov = future_cov.split_after(split)[0] if future_cov else None
-
-        if model_uses_cov:
-            model.fit(
-                train,
-                past_covariates=train_past_cov,
-                future_covariates=train_future_cov,
-            )
-        else:
-            model.fit(train)
-
-    if model_uses_cov:
-        return model.backtest(
-            dataset,
-            retrain=retrain,
-            metric=metric,
-            past_covariates=past_cov,
-            future_covariates=future_cov,
-            start=split,
-            stride=stride,
-        )
-    else:
-        return model.backtest(
-            dataset, retrain=retrain, metric=metric, start=split, stride=stride
-        )
 
 
 metric = mae
